feat(v3_objects): migrate to full dataclass base objects
