+++ src/specklepy/objects_v2/GIS/geometry.py
