+++ src/specklepy/objects_v2/encoding.py
