+++ src/specklepy/objects_v2/graph_traversal/traversal.py
