+++ src/specklepy/objects_v2/structural/analysis.py
