+++ src/specklepy/objects_v2/structural/geometry.py
