+++ src/specklepy/objects_v2/structural/properties.py
