+++ src/specklepy/objects_v2/structural/results.py
